bizhawk: Fix reading outgoing key

+++ crate/multiworld-bizhawk/OotrMultiworld/src/MainForm.cs
@@ -462,7 +462,7 @@ public sealed class MainForm : ToolFormBase, IExternalToolForm {
     private void SendItem(Client client, uint coopContextAddr) {
         ulong outgoingKey;
         if (this.potsanity3) {
-            outgoingKey = APIs.Memory.ReadU32(coopContextAddr + 0x0c1c, "System Bus") << 32;
+            outgoingKey = ((ulong) APIs.Memory.ReadU32(coopContextAddr + 0x0c1c, "System Bus")) << 32;
             outgoingKey |= APIs.Memory.ReadU32(coopContextAddr + 0x0c20, "System Bus");
         } else {
             outgoingKey = APIs.Memory.ReadU32(coopContextAddr + 0xc, "System Bus");
